Popravljeni prošli testovi
