+++ tv_shows_app/src/typings/user/user.ts
@@ -1,0 +1,5 @@
+interface IUser {
+  id: number;
+  email: string;
+  image_url: string;
+}
